description: modify syntax and sub-routine argument data type
file name:
  1. ade7753.cpp
    1. remove sub-routine spi_transfer, modify addr argument data type

  2. ade7753.h
    1. modify addr argument data type

  3. smart_outlet.ino
    1. modify include file name

  4. spi.cpp
    1. delete file

+++ ade7753.h
@@ -52,8 +52,8 @@ class ADE7753 {
     int current;
     int power;
     ADE7753();
-    void wr_reg (int addr, long data, int len);
-    long rd_reg (int addr, int len);
+    void wr_reg (byte addr, long data, int len);
+    long rd_reg (byte addr, int len);
     void rd_irms ();
 };
 #endif
